Docker cleanup and minors

+++ dtest.py
@@ -167,7 +167,7 @@ def running_in_docker():
     return os.path.isfile('/.dockerenv')
 
 
-def cleanup_docker_environment_before_test_execution():
+def cleanup_docker_environment_before_test_execution(reuse_cluster=False):
     """
     perform a bunch of system cleanup operations, like kill any instances that might be
     hanging around incorrectly from a previous run, sync the disk, and clear swap.
@@ -175,9 +175,10 @@ def cleanup_docker_environment_before_test_execution():
     mode there is no way for us to do this.
     """
     # attempt to wack all existing running Cassandra processes forcefully to get us into a clean state
-    p_kill = subprocess.Popen('ps aux | grep -ie CassandraDaemon | grep java | grep -v grep | awk \'{print $2}\' | xargs kill -9',
+    if not reuse_cluster:
+        p_kill = subprocess.Popen('ps aux | grep -ie CassandraDaemon | grep java | grep -v grep | awk \'{print $2}\' | xargs kill -9',
                               stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell=True)
-    p_kill.wait(timeout=10)
+        p_kill.wait(timeout=10)
 
     # explicitly call "sync" to flush everything that might be pending from a previous test
     # so tests are less likely to hit a very slow fsync during the test by starting from a 'known' state
